Cleanup backend/crypto/gpg/cli
Fixes #1880

RELEASE_NOTES=n/a
