--- internal/backend/crypto/gpg/cli/generate_windows_test.go
@@ -1,18 +1,0 @@
-package cli
-
-import (
-	"context"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestGenerateIdentity(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-
-	g := &GPG{}
-	g.binary = "rundll32"
-
-	assert.NoError(t, g.GenerateIdentity(ctx, "foo", "foo@bar.com", "bar"))
-	cancel()
-}
