+++ internal/backend/crypto/gpg/cli/gpg_windows_test.go
@@ -28,3 +28,13 @@ func TestDecrypt(t *testing.T) {
 	assert.NoError(t, err)
 	cancel()
 }
+
+func TestGenerateIdentity(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+
+	g := &GPG{}
+	g.binary = "rundll32"
+
+	assert.NoError(t, g.GenerateIdentity(ctx, "foo", "foo@bar.com", "bar"))
+	cancel()
+}
