+++ internal/backend/crypto/gpg/cli/keyring.go
@@ -44,78 +44,6 @@ func (g *GPG) listKeys(ctx context.Context, typ string, search ...string) (gpg.K
 	return kl, nil
 }
 
-// ListRecipients returns a parsed list of GPG public keys
-func (g *GPG) ListRecipients(ctx context.Context) ([]string, error) {
-	if g.pubKeys == nil {
-		kl, err := g.listKeys(ctx, "public")
-		if err != nil {
-			return nil, err
-		}
-		g.pubKeys = kl
-	}
-	if gpg.IsAlwaysTrust(ctx) {
-		return g.pubKeys.Recipients(), nil
-	}
-	return g.pubKeys.UseableKeys(gpg.IsAlwaysTrust(ctx)).Recipients(), nil
-}
-
-// FindRecipients searches for the given public keys
-func (g *GPG) FindRecipients(ctx context.Context, search ...string) ([]string, error) {
-	kl, err := g.listKeys(ctx, "public", search...)
-	if err != nil || kl == nil {
-		return nil, err
-	}
-
-	recp := kl.UseableKeys(gpg.IsAlwaysTrust(ctx)).Recipients()
-	if gpg.IsAlwaysTrust(ctx) {
-		recp = kl.Recipients()
-	}
-
-	debug.Log("found useable keys for %q: %q (all: %q)", search, recp, kl.Recipients())
-	return recp, nil
-}
-
-// ListIdentities returns a parsed list of GPG secret keys
-func (g *GPG) ListIdentities(ctx context.Context) ([]string, error) {
-	if g.privKeys == nil {
-		kl, err := g.listKeys(ctx, "secret")
-		if err != nil {
-			return nil, err
-		}
-		g.privKeys = kl
-	}
-	if gpg.IsAlwaysTrust(ctx) {
-		return g.privKeys.Recipients(), nil
-	}
-	return g.privKeys.UseableKeys(gpg.IsAlwaysTrust(ctx)).Recipients(), nil
-}
-
-// FindIdentities searches for the given private keys
-func (g *GPG) FindIdentities(ctx context.Context, search ...string) ([]string, error) {
-	kl, err := g.listKeys(ctx, "secret", search...)
-	if err != nil || kl == nil {
-		return nil, err
-	}
-	if gpg.IsAlwaysTrust(ctx) {
-		return kl.Recipients(), nil
-	}
-	return kl.UseableKeys(gpg.IsAlwaysTrust(ctx)).Recipients(), nil
-}
-
-func (g *GPG) findKey(ctx context.Context, id string) gpg.Key {
-	kl, _ := g.listKeys(ctx, "secret", id)
-	if len(kl) >= 1 {
-		return kl[0]
-	}
-	kl, _ = g.listKeys(ctx, "public", id)
-	if len(kl) >= 1 {
-		return kl[0]
-	}
-	return gpg.Key{
-		Fingerprint: id,
-	}
-}
-
 // Fingerprint returns the fingerprint
 func (g *GPG) Fingerprint(ctx context.Context, id string) string {
 	return g.findKey(ctx, id).Fingerprint
