+++ internal/backend/crypto/gpg/gpgconf/binary.go
@@ -1,16 +1,8 @@
-package cli
+package gpgconf
 
 import (
 	"context"
 )
-
-// Binary returns the GPG binary location
-func (g *GPG) Binary() string {
-	if g == nil {
-		return ""
-	}
-	return g.binary
-}
 
 // Binary returns the GPG binary location
 func Binary(ctx context.Context, bin string) (string, error) {
