+++ internal/backend/crypto/gpg/gpgconf/binary_others.go
@@ -1,7 +1,7 @@
 //go:build !windows
 // +build !windows
 
-package cli
+package gpgconf
 
 import (
 	"os/exec"
