+++ internal/backend/crypto/gpg/gpgconf/utils_test.go
@@ -1,4 +1,4 @@
-package cli
+package gpgconf
 
 import (
 	"bytes"
@@ -19,20 +19,6 @@ func TestGpgOpts(t *testing.T) {
 			assert.Equal(t, out, GPGOpts())
 			assert.NoError(t, os.Unsetenv(vn))
 		}
-	}
-}
-
-func TestSplitPacket(t *testing.T) {
-	for in, out := range map[string]map[string]string{
-		"": {},
-		":pubkey enc packet: version 3, algo 1, keyid 00F0FF00FFC00F0F": {
-			"algo":    "1",
-			"keyid":   "00F0FF00FFC00F0F",
-			"version": "3",
-		},
-		":encrypted data packet:": {},
-	} {
-		assert.Equal(t, out, splitPacket(in))
 	}
 }
