+++ internal/backend/crypto/gpg/gpgconf/version_test.go
@@ -1,4 +1,4 @@
-package cli
+package gpgconf
 
 import (
 	"sort"
